requirements.txt
certifi==2020.11.8
chardet==3.0.4
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.4
Pillow==8.0.1
proglog==0.1.9
python-dotenv==0.15.0
pytube==9.7.0
requests==2.24.0
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
